revert 145 and fix writing to existing dir

+++ src/anemoi/utils/testing.py
@@ -156,10 +156,11 @@ def get_test_archive(path: str, extension=".extracted") -> str:
         archive = get_test_data(path)
         target = archive + extension
 
+        if os.path.exists(target):
+            return target
+
         shutil.unpack_archive(archive, os.path.dirname(target) + ".tmp")
         os.rename(os.path.dirname(target) + ".tmp", target)
-
-        os.remove(archive)
 
         return target
